@@ -21,8 +21,8 @@
     "electron": "npx electron .",
     "cp:html": "cpx src/index.html dist",
     "build": "npm-run-all build:*",
-    "build:m": "webpack --config webpack.config.main.js --colors",
-    "build:r": "webpack --config webpack.config.renderer.js --colors",
+    "build:m": "webpack --run-dev --config webpack.main.js",
+    "build:r": "webpack --run-dev --config webpack.renderer.js",
     "watch": "npm-run-all --parallel watch:* watch:cp:*",
     "watch:m": "webpack --config webpack.config.main.js --colors --watch",
     "watch:r": "webpack --config webpack.config.renderer.js --colors --watch",
@@ -30,27 +30,29 @@
     "postinstall": "npm run build"
   },
   "devDependencies": {
-    "@types/node": "^8.0.58",
-    "@types/react": "^16.0.29",
-    "@types/react-dom": "^16.0.2",
-    "awesome-typescript-loader": "^3.4.1",
+    "@types/node": "^8.10.18",
+    "@types/webpack": "^4.1.5",
+    "awesome-typescript-loader": "^5.0.0",
     "cpx": "^1.5.0",
-    "css-loader": "^0.28.7",
-    "electron-packager": "^10.1.0",
-    "file-loader": "^1.1.5",
-    "npm-run-all": "^4.1.2",
-    "semantic-ui-react": "^0.76.0",
-    "style-loader": "^0.19.0",
-    "typescript": "^2.6.1",
-    "uglifyjs-webpack-plugin": "^1.1.4",
-    "webpack": "^3.10.0"
+    "css-loader": "^0.28.11",
+    "del-cli": "^1.1.0",
+    "electron-packager": "^12.1.0",
+    "file-loader": "^1.1.11",
+    "html-loader": "^0.5.5",
+    "node-sass": "^4.9.0",
+    "npm-run-all": "^4.1.3",
+    "postcss": "^6.0.22",
+    "postcss-loader": "^2.1.5",
+    "sass-loader": "^7.0.2",
+    "style-loader": "^0.21.0",
+    "tslint": "^5.10.0",
+    "tslint-loader": "^3.6.0",
+    "typescript": "^2.9.1",
+    "webpack": "^4.10.2",
+    "webpack-command": "^0.2.0"
   },
   "dependencies": {
-    "electron": "^1.7.0",
-    "jschardet": "^1.6.0",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
-    "semantic-ui-css": "^2.2.12",
-    "semantic-ui-react": "^0.77.0"
+    "electron": "^2.0.2",
+    "jschardet": "^1.6.0"
   }
 }
